form validation
a bit of a pain to get svelte form validation to work the way we want it to (late on, early off) without lots of network calls
the preflightOnly makes this possible and so now we have something working but it looks like there are multiple bugs in the sveltekit implementation that is making this difficult.

this is working, so now we just need to make it into a utility function/class

and we might want to alter it slightly so that when server errors are returned (after trying to submit) we keep them displayed until the next submission?