login code input
login code moved to component

lots simpler and nicer

issue: the load function runs when the form returns a value. this results in it seeing a user and redirecting. redirecting from the form action results in it technically “throwing” which will show up as an error.

i could change all errors to returns with success false and a message
or i could try to disambiguate in the catch(e) what the error is and conditionally return based on that.

or i could move that initial page redirect if logged in to the page svelte instead of load function