settings ui
big improvement to how settings are drawn

using bitsui accordion for inline settings “panes” and then will be using probably dialog for the detached mode.

but for accordion, we move the complexity out to an accordion-item and then use that component in settings-content.
So settings content is where you write what settings you want for the app and you use accordion item to create a closed (trigger) for that and then pass a content snippet wrapping the component. so when open, that component will render. and we can pass down a registerAction to link an action button and be able to handle the action button’s click from inside of the component. this way we can encapuslate the complexity away into that component instead of having to polute the parent/wrapper (settings-content)