fix change name issue
previously the settings would break when changing name

due to what svelte labeled a waterfall issue: https://svelte.dev/docs/svelte/runtime-warnings#Client-warnings-await_waterfall

change name needed to make getUser refresh

but the way to access that also needed to be with derived state

for some reason in account-button we need to client side call refresh onMount i think because that component is added/removed from dom based on settings visibility and so if getUser is updated while it’s off screen, then when it comes back it will still have its cached data which is now stale.

there may be better alternatives here like having the data more global and having each of these access it. because each of these places i think is making its own request, so we are duplicating requests of the same data

but this stopped the app from breaking when editing name